requirements.txt
# requirements.txt
Django>=5.0.0,<6.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
dj-rest-auth>=4.0.0
python-decouple>=3.8  # 환경변수 관리
Pillow>=10.0.0  # 이미지 처리용 (프로필/로고 등)

# 개발용 의존성 (선택사항)
# pytest-django>=4.5.0
# black>=23.0.0
# flake8>=6.0.0